+++ modules/apps/commerce/commerce-test/src/testIntegration/java/com/liferay/commerce/order/stock/test/OrderStockManagementTest.java
@@ -36,6 +36,8 @@ import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
 
+import java.math.BigDecimal;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -109,7 +111,7 @@ public class OrderStockManagementTest {
 		CommerceTestUtil.updateBackOrderCPDefinitionInventory(
 			cpInstance.getCPDefinition());
 
-		int orderedQuantity = 4;
+		BigDecimal orderedQuantity = BigDecimal.valueOf(4);
 
 		CommerceOrderItem commerceOrderItem =
 			CommerceTestUtil.addCommerceOrderItem(
@@ -153,7 +155,7 @@ public class OrderStockManagementTest {
 			_commerceChannel.getCommerceChannelId());
 
 		int quantity = 10;
-		int orderedQuantity = 4;
+		BigDecimal orderedQuantity = BigDecimal.valueOf(4);
 
 		CommerceInventoryWarehouseItem commerceInventoryWarehouseItem =
 			CommerceInventoryTestUtil.addCommerceInventoryWarehouseItem(
@@ -191,7 +193,7 @@ public class OrderStockManagementTest {
 
 		Assert.assertEquals(
 			commerceInventoryWarehouseItem.toString(),
-			quantity - orderedQuantity,
+			quantity - orderedQuantity.intValue(),
 			commerceInventoryWarehouseItem.getQuantity());
 	}
 
@@ -226,7 +228,7 @@ public class OrderStockManagementTest {
 			_user.getUserId(), commerceInventoryWarehouse, cpInstance.getSku(),
 			10);
 
-		int orderedQuantity = 2;
+		BigDecimal orderedQuantity = BigDecimal.valueOf(2);
 
 		CommerceOrderItem commerceOrderItem =
 			CommerceTestUtil.addCommerceOrderItem(
@@ -260,7 +262,7 @@ public class OrderStockManagementTest {
 
 		CommerceTestUtil.addCommerceOrderItem(
 			commerceOrder.getCommerceOrderId(), cpInstance.getCPInstanceId(),
-			2);
+			BigDecimal.valueOf(2));
 	}
 
 	@Test(expected = CommerceOrderValidatorException.class)
@@ -300,7 +302,7 @@ public class OrderStockManagementTest {
 
 		CommerceTestUtil.addCommerceOrderItem(
 			commerceOrder.getCommerceOrderId(), cpInstance.getCPInstanceId(),
-			20);
+			BigDecimal.valueOf(20));
 	}
 
 	@Test(expected = CommerceOrderValidatorException.class)
@@ -347,7 +349,7 @@ public class OrderStockManagementTest {
 
 		CommerceTestUtil.addCommerceOrderItem(
 			commerceOrder1.getCommerceOrderId(), cpInstance.getCPInstanceId(),
-			4);
+			BigDecimal.valueOf(4));
 
 		CommerceShipmentTestUtil.createOrderShipment(
 			_user.getGroupId(), commerceOrder1.getCommerceOrderId(),
@@ -355,7 +357,7 @@ public class OrderStockManagementTest {
 
 		CommerceTestUtil.addCommerceOrderItem(
 			commerceOrder2.getCommerceOrderId(), cpInstance.getCPInstanceId(),
-			8);
+			BigDecimal.valueOf(8));
 	}
 
 	@Rule
